requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0
numpy>=1.20.0
pyyaml>=6.0
markdown>=3.4.0
mistune>=2.0.0
pygments>=2.10.0
requests>=2.25.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
scipy>=1.7.0
plotly>=5.3.0
prophet>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
nltk>=3.6.0
gensim>=4.1.0
networkx>=2.6.0
